Move RibbonCategoryLayoutWidget to category.py
